naive-rag-system/pyproject.toml
[project]
name = "naive-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langchain-postgres>=0.0.17",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "unstructured-inference>=1.5.3",
    "unstructured[local-inference,pdf]>=0.22.6",
    "uvicorn>=0.42.0",
]
